+++ tests/train/conftest.py
@@ -26,7 +26,7 @@ from typing import Union
 import numpy as np
 from pytest import fixture
 
-from deepdoctection.dataflow import DataFlow, DataFromList  # type: ignore
+from deepdoctection.dataflow import DataFlow, DataFromList
 from deepdoctection.datapoint import BoundingBox, Image, ImageAnnotation
 from deepdoctection.datasets.base import DatasetBase
 from deepdoctection.datasets.dataflow_builder import DataFlowBaseBuilder
@@ -47,7 +47,7 @@ class Datapoint:
     """
 
     image = Image(location="/test/to/path", file_name="test_name")
-    np_image = np.ones([400, 600, 3], dtype=np.float32)
+    np_image = np.ones([400, 600, 3], dtype=np.uint8)
     anns = [
         ImageAnnotation(
             bounding_box=BoundingBox(ulx=100.0, uly=100.0, lrx=200.0, lry=400.0, absolute_coords=True),
